scripts/requirements.txt
anthropic
resend
google-auth
google-api-python-client
